Refactor volume sliders a bit
